requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==20.1.0
astroid==2.3.1
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
bleach==3.3.1
bokeh==2.4.0
branca==0.5.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.1
coloredlogs==15.0.1
constantly==15.1.0
cryptography==36.0.2
cycler==0.10.0
dask==2021.12.0
datascience==0.17.2
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
folium==0.12.1.post1
fonttools==4.28.2
frozenlist==1.3.0
fsspec==2022.3.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
htmlmin==0.1.12
httplib2==0.20.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.1
ImageHash==4.2.1
imbalanced-learn==0.9.0
imblearn==0.0
img2pdf==0.4.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
incremental==17.5.0
inflect==5.4.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpath-ng==1.5.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kiwisolver==1.3.1
lazy-object-proxy==1.4.2
locket==0.2.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
mccabe==0.6.1
Metaphone==0.6
missingno==0.5.0
mistune==0.8.4
multidict==6.0.2
multimethod==1.6
mysql-connector-python==8.0.28
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.7
notebook==6.4.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.2.0
ocrmypdf==13.4.1
opencv-python==4.5.5.64
opendatasets==0.1.20
openpyxl==3.0.7
packaging==21.0
pandas==1.3.4
pandas-profiling==3.1.0
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pdfminer.six==20211012
pdfplumber==0.6.0
phik==0.12.0
pickleshare==0.7.5
pikepdf==5.1.0
Pillow==9.0.1
pipenv==2022.5.2
platformdirs==2.5.2
plotly==5.3.1
pluggy==1.0.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.19.4
pure-eval==0.2.2
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyDrive==1.3.1
Pygments==2.9.0
PyHamcrest==2.0.2
pylint==2.4.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyreadline==2.1
pyrsistent==0.18.0
pyspark==3.1.2
pystache==0.6.0
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-slugify==5.0.2
python-stdnum==1.17
pytz==2021.1
PyWavelets==1.2.0
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.11.10
reportlab==3.6.9
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.7.1
setuptools-scm==6.3.2
six==1.12.0
SQLAlchemy==1.4.35
sweetviz==2.1.3
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tomli==1.2.2
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
Twisted @ file:///C:/Users/Naveen/Downloads/Twisted-20.3.0-cp37-cp37m-win_amd64.whl
typed-ast==1.4.0
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.7
varname==0.8.3
virtualenv==20.14.1
virtualenv-clone==0.5.7
visions==0.7.4
Wand==0.6.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.11.2
xgboost==1.6.1
yarl==1.7.2
zipp==3.5.0
zope.interface==5.2.0
